requirements.txt
attrs==19.3.0
beautifulsoup4==4.9.1
hupper==1.10.2
more-itertools==8.4.0
packaging==20.4
PasteDeploy==2.1.0
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.13.1
py==1.8.2
pyparsing==2.4.7
pyramid==1.10.4
pytest==5.4.3
simplecalculator==0.0.4
six==1.15.0
soupsieve==2.0.1
translationstring==1.3
venusian==3.0.0
waitress==1.4.4
wcwidth==0.2.4
WebOb==1.8.6
WebTest==2.0.35
zope.deprecation==4.4.0
zope.interface==5.1.0
